reorganize structure code in quantstudio design
