+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/appbio_quantstudio_designandanalysis_data_creator.py
@@ -5,12 +5,20 @@ from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_d
 from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_structure import (
     Data,
 )
-from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.creators import (
-    genotyping,
-    melt_curve,
-    presence_absence,
-    relative_standard_curve,
-    standard_curve,
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.genotyping.creator import (
+    GenotypingCreator,
+)
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.melt_curve.creator import (
+    MeltCurveCreator,
+)
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.presence_absence.creator import (
+    PresenceAbsenceCreator,
+)
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.relative_standard_curve.creator import (
+    RelativeStandardCurveCreator,
+)
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.standard_curve.creator import (
+    StandardCurveCreator,
 )
 
 
@@ -18,11 +26,11 @@ def create_data(contents: DesignQuantstudioContents) -> Data:
     possible_creators = [
         creator
         for creator in [
-            standard_curve.StandardCurveCreator,
-            relative_standard_curve.RelativeStandardCurveCreator,
-            genotyping.GenotypingCreator,
-            melt_curve.MeltCurveCreator,
-            presence_absence.PresenceAbsenceCreator,
+            StandardCurveCreator,
+            RelativeStandardCurveCreator,
+            GenotypingCreator,
+            MeltCurveCreator,
+            PresenceAbsenceCreator,
         ]
         if creator.check_type(contents)
     ]
