+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/structure/standard_curve/creator.py
@@ -14,10 +14,10 @@ from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_d
     TargetRoleView,
     TargetView,
 )
-from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.standard_curve import (
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.standard_curve.structure import (
     StandardCurveWellList,
 )
-from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.creators.generic import (
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.generic.creator import (
     Creator,
 )
