+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/structure/standard_curve/structure.py
